paso de GET y DELETE de routes a index.js, eliminacion de routes, se agrega PUT a index.js, habilitacion de nodemon
