+++ index.html
@@ -18,19 +18,19 @@
         <div class="form-group row">
           <label for="name" class="col-sm-2 col-form-label">Canción:</label>
           <div class="col-sm-10">
-            <input type="text" class="form-control" id="cancion" value="A dios le pido" />
+            <input type="text" class="form-control" id="cancion" value="" />
           </div>
         </div>
         <div class="form-group row">
           <label for="email" class="col-sm-2 col-form-label">Artista: </label>
           <div class="col-sm-10">
-            <input type="text" class="form-control" id="artista" value="Juanes" />
+            <input type="text" class="form-control" id="artista" value="" />
           </div>
         </div>
         <div class="form-group row">
           <label for="rut" class="col-sm-2 col-form-label">Tono:</label>
           <div class="col-sm-10">
-            <input type="text" class="form-control" id="tono" value="Em" />
+            <input type="text" class="form-control" id="tono" value="" />
           </div>
         </div>
         <button onclick="nuevaCancion()" id="agregar" class="m-auto btn btn-success">
