+++ package.json
@@ -5,13 +5,12 @@
    "main": "index.js",
    "scripts": {
       "test": "echo \"Error: no test specified\" && exit 1",
-      "start": "node index.js"
+      "start": "nodemon index.js"
    },
    "keywords": [],
-   "author": "Mónica Ayala",
-   "license": "ISC",
+   "author": "Mónica Ayala - Ricardo Barrientos",
+   "license": "MIT",
    "dependencies": {
-      "cors": "^2.8.5",
       "express": "^4.18.2"
    },
    "devDependencies": {
